update to use httpexceptor
Also move the imports out to the top level of the module,
as this makes import errors show up much sooner.

please merge and release @fnd?

Due to the lack of tests I've tested it in a dev to insure that
it loads and operates in it's basic exploring form. It does.