change twp.console to use imaker and ibuilder
Quite a lot of changes here, but there is rationale for them.
Review from @FND desired but not required. If/when this gets
merged it will need to be synchronized with i{maker,builder}
changes in tiddlywebwiki and tiddlyspace.

* Get rid of standalone application build process.
* Get rid of tiddlywiki macro/plugin stuff.
* Get rid of cacher in favor of calling twibuilder.
* Put .js.js extension on tiddlers which will be loaded as *.js.
* Add twconsole instance creation script (useful for testing).

I tried to update to modern chrjs and modern jquery and failed
miserably, so backed off to the static version that are listed
in the Makefile. The chrjs API changed a lot after this code
was created. Even with the old chrjs in place new jquery would
not work and I couldn't be bothered to investigate as I generally
consider this code legacy, in the sense that it is useful but
not really worth making better. A modern version would have a
quite different approach based all we've learned in the past
few years.

That legacy-ness is what drives the removal of the standalone
handling and the the tiddlywiki macro handling. While these
may have been useful at one point in the past the primary
purpose these days of twp.console is to provide /console on a
running tiddlywebwiki instance.

Getting rid of that stuff means we can use default twimport
handling when processing tiddlers into the packages resources,
we don't need to munge paths, etc. So we end up not needing
the special cacher script, nor the complex instance.py

The end result provides the same functionality with less
noise and less fuss. It's sort of annoying or disappointing
to remove the other functionality and kill the TODO stuff,
but it's not really all that relevant anymore. Or is it?